requirements.txt
streamlit
pillow
pytesseract
fuzzywuzzy
python-Levenshtein

